Not an integration test. Fix test.
Moved the test in `tests/integration/utils/test_thin.py` to `tests/unit/utils/test_thin.py`.
Additionally, fixed the test to use the `VirtualEnv` helper instead of
the virtualenv module directly.

Fixes #57386
